코드 정리 및 로딩 컴포넌트 추가
